requirements.txt
alembic==1.4.0
aniso8601==8.1.1
autopep8==1.5.5
cffi==1.14.0
Click==7.0
cryptography==2.8
dominate==2.5.1
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.10
Flask-HTTPAuth==3.3.0
Flask-Migrate==2.5.2
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.1
linecache2==1.0.0
Mako==1.1.1
MarkupSafe==1.1.1
passlib==1.7.2
psycopg2==2.8.4
psycopg2-binary==2.8.4
pycodestyle==2.6.0
pycparser==2.19
PyJWT==1.7.1
pyodbc==4.0.30
python-dateutil==2.8.1
python-dotenv==0.10.5
python-editor==1.0.4
pytz==2021.1
six==1.14.0
SQLAlchemy==1.3.13
toml==0.10.2
traceback2==1.4.0
unittest2==1.1.0
visitor==0.1.3
Werkzeug==0.16.1
